[Acceloremeter] Creation of base class
This class implements the driver for the LIS3DSH accelerometer.
The class inherits from Device, it will be registered to the Devfs
pseudo filesystem and exposed through the client via a file.
By means of this file, one can write to it and select the axis,
and read the data collected by the sensor from the previous specified axis.

Signed-off-by: Raffaele Fioratto <raffaele.fioratto@gmail.com>