[Accelerometer] Implemented driver's constructor
The configuration of the accelerometer is carried out in the constructor.
After initializing the GPIOs and the SPI master, the code checks for the actual
presence of the LIS3DSH accelerometer and performs initial configuration.
By default, the selected axis from which the reads are perfomed is the X axis.

Signed-off-by: Raffaele Fioratto <raffaele.fioratto@gmail.com>