[Accelerometer] Driver code implementation
This code implement the Device interface.
The result is a functional driver that is ready to be registered to the Devfs
pseudo filesystem.
By writing to the file one can select the axis between 0 and 2 (inclusive).
By reading, the data acquired by the sensor of the previously selected axis is returned
to the client application.

Signed-off-by: Raffaele Fioratto <raffaele.fioratto@gmail.com>