[Accelerometer] Code for registering LIS3DSH Driver to Devfs
This codes performs the actual registration of the LIS3DSH Driver during bootup
of the stm32f4 board. The code is compiled only when the filesystem implementation is
included in the kernel (which is by default).
During boot a message is printed through the serial port indicating whether the
initialization succeded or not.

Signed-off-by: Raffaele Fioratto <raffaele.fioratto@gmail.com>