[pendulum] Use float variabled instead of double.
This change has the effect of improving code size, memory usage at runtime and also performances.
For this board, the operations with variables of type float are done in hardware (fast) whereas
with double, they are implemented in software (slow). Moreover, the extra precision of doubles
is not relevant in our context so we can safely fallback to floats.

Signed-off-by: Raffaele Fioratto <raffaele.fioratto@gmail.com>